+++ DemoApps/Vulkan/Bloom/source/Bloom.cpp
@@ -128,15 +128,13 @@ namespace Fsl
       std::array<VkWriteDescriptorSet, 1> writeDescriptorSets{};
 
       // Binding 0 : Fragment shader texture sampler
-      {
-        auto textureImageInfo = texture.GetDescriptorImageInfo();
-        writeDescriptorSets[0].sType = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET;
-        writeDescriptorSets[0].dstSet = descriptorSet;
-        writeDescriptorSets[0].dstBinding = 0;
-        writeDescriptorSets[0].descriptorCount = 1;
-        writeDescriptorSets[0].descriptorType = VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER;
-        writeDescriptorSets[0].pImageInfo = &textureImageInfo;
-      }
+      auto textureImageInfo0 = texture.GetDescriptorImageInfo();
+      writeDescriptorSets[0].sType = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET;
+      writeDescriptorSets[0].dstSet = descriptorSet;
+      writeDescriptorSets[0].dstBinding = 0;
+      writeDescriptorSets[0].descriptorCount = 1;
+      writeDescriptorSets[0].descriptorType = VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER;
+      writeDescriptorSets[0].pImageInfo = &textureImageInfo0;
 
       vkUpdateDescriptorSets(device, static_cast<uint32_t>(writeDescriptorSets.size()), writeDescriptorSets.data(), 0, nullptr);
       return descriptorSet;