+++ DemoApps/Vulkan/T3DStressTest/source/MeshRenderVBInstanced.cpp
@@ -391,8 +391,11 @@ namespace Fsl
 
   void MeshRenderVBInstanced::OnBuildResources(const VulkanBasic::BuildResourcesContext& context, const VkRenderPass renderPass)
   {
+    FSLLOG3_VERBOSE3("MeshRenderVBInstanced::OnBuildResources");
+    FSLLOG3_VERBOSE3("- Creating opaque pipeline");
     m_dependentResources.OpaquePipeline = CreatePipeline(m_resources.PipelineLayout, context.SwapchainImageExtent, VertShader.Get(), FragShader.Get(),
                                                          m_resources.VB, renderPass, m_topology, true, m_enableDepthTest);
+    FSLLOG3_VERBOSE3("- Creating pipeline");
     m_dependentResources.Pipeline = CreatePipeline(m_resources.PipelineLayout, context.SwapchainImageExtent, VertShader.Get(), FragShader.Get(),
                                                    m_resources.VB, renderPass, m_topology, false, m_enableDepthTest);
   }