+++ DemoFramework/FslUtil/Vulkan1_0/include/FslUtil/Vulkan1_0/Util/PhysicalDeviceUtil.hpp
@@ -112,13 +112,19 @@ namespace Fsl
       }
 
       //! @brief Find a VK_IMAGE_TILING_OPTIMAL depth format
-      inline VkFormat FindDepthFormat(const VkPhysicalDevice physicalDevice)
+      inline VkFormat FindDepthFormat(const VkPhysicalDevice physicalDevice, const bool enableStencil)
       {
         assert(physicalDevice != VK_NULL_HANDLE);
-        std::array<VkFormat, 5> depthFormats = {VK_FORMAT_D32_SFLOAT, VK_FORMAT_D32_SFLOAT_S8_UINT, VK_FORMAT_D24_UNORM_S8_UINT,
-                                                VK_FORMAT_D16_UNORM_S8_UINT, VK_FORMAT_D16_UNORM};
+        if (!enableStencil)
+        {
+          constexpr std::array<VkFormat, 2> depthFormats = {VK_FORMAT_D32_SFLOAT, VK_FORMAT_D16_UNORM};
+          return PhysicalDeviceUtil::FindSupportedFormat(physicalDevice, depthFormats, VK_IMAGE_TILING_OPTIMAL,
+                                                         VK_FORMAT_FEATURE_DEPTH_STENCIL_ATTACHMENT_BIT);
+        }
 
-        return PhysicalDeviceUtil::FindSupportedFormat(physicalDevice, depthFormats, VK_IMAGE_TILING_OPTIMAL,
+        constexpr std::array<VkFormat, 3> depthFormatsWithS = {VK_FORMAT_D32_SFLOAT_S8_UINT, VK_FORMAT_D24_UNORM_S8_UINT,
+                                                               VK_FORMAT_D16_UNORM_S8_UINT};
+        return PhysicalDeviceUtil::FindSupportedFormat(physicalDevice, depthFormatsWithS, VK_IMAGE_TILING_OPTIMAL,
                                                        VK_FORMAT_FEATURE_DEPTH_STENCIL_ATTACHMENT_BIT);
       }
     }